+++ src/components/LoginForm/index.css
@@ -1,0 +1,9 @@
+.loginForm {
+  max-width: 300px;
+}
+.loginFormForgot {
+  float: right;
+}
+.loginFormButton {
+  width: 100%;
+}
