+++ src/components/LoginForm/index.js
@@ -1,0 +1,63 @@
+import { Component } from "react";
+import { Form, Icon, Input, Button, Checkbox } from "antd";
+import styles from "./index.css";
+const FormItem = Form.Item;
+
+class LoginForm extends Component {
+  handleSubmit = e => {
+    e.preventDefault();
+    this.props.form.validateFields((err, values) => {
+      if (!err) {
+        console.log("Received values of form: ", values);
+      }
+    });
+  };
+
+  render() {
+    const { getFieldDecorator } = this.props.form;
+    return (
+      <Form onSubmit={this.handleSubmit} className={styles.loginForm}>
+        <FormItem>
+          {getFieldDecorator("userName", {
+            rules: [{ required: true, message: "Please input your username!" }]
+          })(
+            <Input
+              prefix={<Icon type="user" style={{ color: "rgba(0,0,0,.25)" }} />}
+              placeholder="Username"
+            />
+          )}
+        </FormItem>
+        <FormItem>
+          {getFieldDecorator("password", {
+            rules: [{ required: true, message: "Please input your Password!" }]
+          })(
+            <Input
+              prefix={<Icon type="lock" style={{ color: "rgba(0,0,0,.25)" }} />}
+              type="password"
+              placeholder="Password"
+            />
+          )}
+        </FormItem>
+        <FormItem>
+          {getFieldDecorator("remember", {
+            valuePropName: "checked",
+            initialValue: true
+          })(<Checkbox>Remember me</Checkbox>)}
+          <a className={styles.loginFormForget} href="">
+            Forgot password
+          </a>
+          <Button
+            type="primary"
+            htmlType="submit"
+            className={styles.loginFormButton}
+          >
+            Log in
+          </Button>
+          Or <a href="">register now!</a>
+        </FormItem>
+      </Form>
+    );
+  }
+}
+
+export default Form.create()(LoginForm);
